Rationalized the versioning and package information
The author, URL, etc are now expressed in a file embedded in the package,
__about__.py. This can be loaded in setup.py and its slots used as part of
the build and test process. The result is a single source of truth for the
package, no matter how it is consumed or used. The onyl caveat is that
`import toplevelmodule.__about__` must be possible before any of dependencies
specified in `requirements.txt` are installed.