Fixes to doc source and Python build support
Made a few ReST syntax edits to source files to resolve warnings/errors during Sphinx builds. Updated the
``conf.py`` file to leverage the single-source-of-truth version and package metadata. It is also updated
to use relative or submodule imports in constructing tables and such in the documentation. The build log
is still a fantasy of light, sound, and warnings though.